Move all relevant tests to orders app
